Fix node build
Presumably due to it being exported in two places, `loadFromJSON` was
mangled to `loadFromJSON$1`.  This was exported correctly, *however*
`loadFromFile`'s body was not updated to use the mangled name.

I'm not sure the source of the problem: I suspect it's somewhere in
rollup core.  For now, manually wrapping the reexports works around it.